requirements.txt
openai==1.13.3
langchain==0.1.11
langchain-openai==0.0.8
loguru==0.7.2
sentence-transformers==2.2.2
unstructured==0.6.6
pydantic==1.10.11
typing-inspect==0.8.0
typing_extensions==4.10.0
chromadb==0.3.26
fastapi==0.110.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
python-dotenv ==0.20.0
SQLAlchemy==2.0.0
starlette==0.36.3
